refactored notify plugin and removed unused variables
